Added lots of fixes and additions (buttons for delete/edit of products)
Many smaller fixes done.
Added buttons for delete/edit of products and hooked up to the functions
that processes them on POST. The functions need more work.
Edit button works (gives values to the form for editing, saving the
values not done since that is part of the whole save function)