Validation logic is done, with a couple of issues left. Started on logic for differentiating between new product, edit products cases.
Issues1:
When a file with too large size it submittet, an error message comes
wordpress or the server, bypassing all out code. It also resets the
whole POST.

Issue 2:
The validation for the image type might not work completely as it
should. I have one image that it complains about, that i can't see why
it complains.